Decouple the OAuth methods from the data processing methods
The User class now solely manages the information needed to interact
with the Spotify API. The functions for actually fetching and
processing the data from the API are now in the client_services
module. This decoupling should reduce the complexity in making
future changes, and separation of concerns is more clear. The
flask endpoints now accordingly retrieve the client from the
User object, and pass it to the client_services functions.
Checking for token expiry is now simplified - we only need to check
when the client is retrieved, not before each call to the API.