Use flask-dance for OAuth flow and token management
It turns out there was an existing solution to handle OAuth and
the tokens, so switching to that instead of the existing User
class. This simplifies the session management a lot, and it's also
easy enough to automate initiating the OAuth process if the user
doesn't have a token, or if the token expired.