Change endpoint naming for playlist interactions
Slight change to hierarchy - it seems more intuitive to move some
of the endpoints related to retrieving the list of user playlists
and creating playlists to the /user/ hierarchy. These endpoints
were changed specifically because they're operating on behalf of
the user, and only concern the playlists which the user controls.
May move the other playlist related endpoints later, or move these
back out depending on how easily understandable the distinction
actually is.