Add playlist pages/components
Implemented pages for listing the user's playlists as well as
for individual playlists, along with a component to display the
averaged audio attributes of the playlist. The chart used in this
component requires the data to be formatted in a specific way, so
the backend was also updated to handle this accordingly.