Add basic error handling for 404 and 500 responses
Implemented simple error handlers in the backend to return proper
error codes. Also added components for the front-end to display
when an error is encountered; one for error responses it receives
from the backend, and one for bad Route URLs.
The most likely errors would be from users entering
URLs manually rather than navigating via page links, which is
reflected in the 500 response page.